Tweak config handling and add output of untagged movies to radarr
The !radarr command now lists all untagged movies.

Config is now loaded locally rather than into the environment. Additionally,
if a DISCORD_CHANNEL is listed then it only responds to messages on that
channel. This is for testing purposes so that a second bot can be tested
without having to restart the first one.